// src/types/materialindisponivel.ts
export type MaterialIndisponivel = {
  cmdo: string;
  rm: string;
  de: string;
  bda: string;
  om: string;
  idsigelog: string;
  equipamento: string;
  modelo: string;
  fabricante: string;
  pn: string;
  sn: string;
  motivoindisp: string;
  tipoEqp: string;
  subsistema: string;
  grupo: string;
};